Fixes:
Build system:
* Android: fixed external static libraries export to Android.mk

Common:
* Time - fixed constexpr

Layout:
* Added TextDecoration::[LineThrough, Overline]

Stappler:
* ScrollController: refactored update function
* ScrollItemHandle component allows to define scroll items without
height precalculations

+++ stappler/src/features/storekit/SPStoreKit.cc
@@ -155,7 +155,7 @@ bool StoreKit::isSubscriptionActive(const std::string &productId) {
 	if (auto product = getProduct(productId)) {
 		if (product->isSubscription && product->expiration) {
 			auto now = Time::now();
-			if (product->expiration > (int64_t)now) {
+			if (product->expiration > now) {
 				return true;
 			}
 		}
